requirements.txt
paho-mqtt
gpiozero
adafruit-circuitpython-dht
adafruit-blinka
